requirements.txt
# Web framework
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
httpx==0.27.0

# Reddit API
praw==7.7.1

# Database
psycopg2-binary==2.9.9

# Redis
redis==5.0.3

# Vector search
faiss-cpu==1.13.2

# Embeddings
sentence-transformers==2.6.1
numpy==1.26.4

# HTTP / LLM
requests==2.31.0

# Config
python-dotenv==1.0.1

# Data validation
pydantic==2.6.3
pydantic-settings==2.2.1
